Added Case Reset
Clicking reset case will reset the case with the original thickness. User call also reclick the generate casing button to recreate the case with new thickness, but this will remove the cuts. I tried to change thickness with the cuts, but it seems to be quite tricky, so I have this for now.